Time: 26 ms (13.73%), Space: 48.1 MB (29.97%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -23,7 +23,7 @@ class Solution {
     private boolean kokoCanDoIt (int[] piles, int speed, int h) {
         int sum = 0;
         for (int i = 0; i < piles.length; i++) {
-            sum += (int) Math.ceil((double)piles[i]/speed);
+            sum += Math.ceil((double)piles[i]/speed);
 
             if (sum > h) return false;
         }
